refactor : extract Bio into a single component
